Extract RootParser from Ingreedy::Parser
Ingreedy::Parser grew to do enough outside of the parslet DSl that I
found it more readable and maintainable with a separate RootParser
class.
